Make getRelatedIconClass separate util, refactor Section component API
Section now has a prop called icon which you can pass in an icon class
(e.g. octicon-database) and we automatically concatenate it with the
layout classname inside the component. This way a user does not have to
be aware of said classname.
