+++ admin/client/App/screens/Home/components/Section.js
@@ -1,39 +1,13 @@
 import React from 'react';
+import getRelatedIconClass from '../utils/getRelatedIconClass';
 
 class Section extends React.Component {
-	// Certain section name have an icon associated with the for a nicer view
-	getHeadingIconClasses (navSectionKey) {
-		const icons = [
-			{ icon: 'book', sections: ['books', 'posts', 'blog', 'blog-posts', 'stories', 'news-stories', 'content'] },
-			{ icon: 'briefcase', sections: ['businesses', 'companies', 'listings', 'organizations', 'partners'] },
-			{ icon: 'calendar', sections: ['events', 'dates'] },
-			{ icon: 'clock', sections: ['classes', 'hours', 'times'] },
-			{ icon: 'file-media', sections: ['gallery', 'galleries', 'images', 'photos', 'pictures'] },
-			{ icon: 'file-text', sections: ['attachments', 'docs', 'documents', 'files'] },
-			{ icon: 'location', sections: ['locations', 'markers', 'places'] },
-			{ icon: 'mail', sections: ['emails', 'enquiries'] },
-			{ icon: 'megaphone', sections: ['broadcasts', 'jobs', 'talks'] },
-			{ icon: 'organization', sections: ['contacts', 'customers', 'groups', 'members', 'people', 'speakers', 'teams', 'users'] },
-			{ icon: 'package', sections: ['boxes', 'items', 'packages', 'parcels'] },
-			{ icon: 'tag', sections: ['tags'] },
-		];
-		const classes = icons
-			.filter(obj => obj.sections.indexOf(navSectionKey) !== -1)
-			.map(obj => `octicon-${obj.icon}`);
-
-		if (!classes.length) {
-			classes.push('octicon-primitive-dot');
-		}
-
-		return ['dashboard-group__heading-icon', 'octicon', ...classes].join(' ');
-	}
-
 	render () {
-		const iconClassname = this.props.iconClass || this.getHeadingIconClasses(this.props.id);
+		const iconClass = this.props.icon || getRelatedIconClass(this.props.id);
 		return (
 			<div className="dashboard-group" data-section-label={this.props.label}>
 				<div className="dashboard-group__heading">
-					<span className={iconClassname} />
+					<span className={`dashboard-group__heading-icon octicon ${iconClass}`} />
 					{this.props.label}
 				</div>
 				{this.props.children}
@@ -44,7 +18,7 @@ class Section extends React.Component {
 
 Section.propTypes = {
 	children: React.PropTypes.element.isRequired,
-	iconClass: React.PropTypes.string,
+	icon: React.PropTypes.string,
 	id: React.PropTypes.string,
 	label: React.PropTypes.string.isRequired,
 };
